// FRONT/RECIBOS/src/app/models/Articulo.ts
export interface FiltrosArt {
  filtro: string;
  TipoProducto: string;
  Resistencia: string;
  Suaje: string;
  idCliente: string;
  generico: boolean;
  sinBackFrom: boolean;
  Suspendido: boolean;
  Estatus: boolean;
  ComponentePrin?: string;
}